move : 펫시터 관련 파일 펫시터 도메인 폴더로 이동
