+++ src/main/java/com/ajou_nice/with_pet/domain/entity/ReservationPetSitterService.java
@@ -8,6 +8,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterWithPetService;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
