+++ src/main/java/com/ajou_nice/with_pet/petsitter/model/dto/HouseInfoResponse.java
@@ -1,7 +1,7 @@
-package com.ajou_nice.with_pet.domain.dto.petsitter;
+package com.ajou_nice.with_pet.petsitter.model.dto;
 
 
-import com.ajou_nice.with_pet.domain.entity.House;
+import com.ajou_nice.with_pet.petsitter.model.entity.House;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
