+++ src/main/java/com/ajou_nice/with_pet/petsitter/model/entity/PetSitterHashTag.java
@@ -1,7 +1,7 @@
-package com.ajou_nice.with_pet.domain.entity;
+package com.ajou_nice.with_pet.petsitter.model.entity;
 
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterHashTagRequest;
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterHashTagRequest;
+
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -10,11 +10,11 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
 @Entity
