+++ src/main/java/com/ajou_nice/with_pet/petsitter/repository/PetSitterCriticalServiceRepository.java
@@ -1,7 +1,7 @@
-package com.ajou_nice.with_pet.repository;
+package com.ajou_nice.with_pet.petsitter.repository;
 
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
-import com.ajou_nice.with_pet.domain.entity.PetSitterCriticalService;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterCriticalService;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
