+++ src/main/java/com/ajou_nice/with_pet/petsitter/repository/PetSitterHashTagRepository.java
@@ -1,9 +1,8 @@
-package com.ajou_nice.with_pet.repository;
+package com.ajou_nice.with_pet.petsitter.repository;
 
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
-import com.ajou_nice.with_pet.domain.entity.PetSitterHashTag;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterHashTag;
 import java.util.List;
-import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
