+++ src/main/java/com/ajou_nice/with_pet/petsitter/repository/custom/PetSitterRespositoryCustom.java
@@ -1,6 +1,6 @@
-package com.ajou_nice.with_pet.repository.custom;
+package com.ajou_nice.with_pet.petsitter.repository.custom;
 
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
