+++ src/main/java/com/ajou_nice/with_pet/petsitter/service/PetSitterService.java
@@ -1,34 +1,35 @@
-package com.ajou_nice.with_pet.service;
+package com.ajou_nice.with_pet.petsitter.service;
 
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterDetailInfoResponse;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterDetailInfoResponse.PetSitterModifyInfoResponse;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterMainResponse;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterCriticalServiceRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterCriticalServicesRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterHashTagRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterHashTagsRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterHouseRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterHousesRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterInfoRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterIntroRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterServiceRequest;
-import com.ajou_nice.with_pet.domain.dto.petsitter.PetSitterRequest.PetSitterWithPetServicesRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterDetailInfoResponse;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterDetailInfoResponse.PetSitterModifyInfoResponse;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterMainResponse;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterCriticalServiceRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterCriticalServicesRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterHashTagRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterHashTagsRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterHouseRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterHousesRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterInfoRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterIntroRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterServiceRequest;
+import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterRequest.PetSitterWithPetServicesRequest;
 import com.ajou_nice.with_pet.critical_service.model.entity.CriticalService;
-import com.ajou_nice.with_pet.domain.entity.House;
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
-import com.ajou_nice.with_pet.domain.entity.PetSitterCriticalService;
-import com.ajou_nice.with_pet.domain.entity.PetSitterHashTag;
-import com.ajou_nice.with_pet.domain.entity.PetSitterWithPetService;
+import com.ajou_nice.with_pet.petsitter.model.entity.House;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterCriticalService;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterHashTag;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitterWithPetService;
 import com.ajou_nice.with_pet.domain.entity.Review;
 import com.ajou_nice.with_pet.domain.entity.User;
+import com.ajou_nice.with_pet.service.ValidateCollection;
 import com.ajou_nice.with_pet.withpet_service.model.entity.WithPetService;
 import com.ajou_nice.with_pet.enums.DogSize;
 import com.ajou_nice.with_pet.critical_service.repository.CriticalServiceRepository;
-import com.ajou_nice.with_pet.repository.HouseRepository;
-import com.ajou_nice.with_pet.repository.PetSitterCriticalServiceRepository;
-import com.ajou_nice.with_pet.repository.PetSitterHashTagRepository;
-import com.ajou_nice.with_pet.repository.PetSitterRepository;
-import com.ajou_nice.with_pet.repository.PetSitterServiceRepository;
+import com.ajou_nice.with_pet.petsitter.repository.HouseRepository;
+import com.ajou_nice.with_pet.petsitter.repository.PetSitterCriticalServiceRepository;
+import com.ajou_nice.with_pet.petsitter.repository.PetSitterHashTagRepository;
+import com.ajou_nice.with_pet.petsitter.repository.PetSitterRepository;
+import com.ajou_nice.with_pet.petsitter.repository.PetSitterServiceRepository;
 import com.ajou_nice.with_pet.repository.ReviewRepository;
 import com.ajou_nice.with_pet.withpet_service.repository.WithPetServiceRepository;
 import java.util.ArrayList;
