+++ src/main/java/com/ajou_nice/with_pet/repository/DiaryRepository.java
@@ -2,7 +2,7 @@ package com.ajou_nice.with_pet.repository;
 
 import com.ajou_nice.with_pet.domain.entity.Diary;
 import com.ajou_nice.with_pet.domain.entity.Dog;
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
 import com.ajou_nice.with_pet.repository.custom.userdiary.DiaryRepositoryCustom;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
