+++ src/main/java/com/ajou_nice/with_pet/repository/custom/reservation/ReservationRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.ajou_nice.with_pet.repository.custom.reservation;
 
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
 import com.ajou_nice.with_pet.domain.entity.Reservation;
 import com.ajou_nice.with_pet.enums.ReservationStatus;
 import java.time.LocalDate;
