+++ src/main/java/com/ajou_nice/with_pet/repository/custom/reservation/ReservationRepositoryImpl.java
@@ -2,7 +2,7 @@ package com.ajou_nice.with_pet.repository.custom.reservation;
 
 import static com.ajou_nice.with_pet.domain.entity.QReservation.reservation;
 
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
 import com.ajou_nice.with_pet.domain.entity.Reservation;
 import com.ajou_nice.with_pet.enums.ReservationStatus;
 import com.querydsl.core.types.dsl.BooleanExpression;
