+++ src/main/java/com/ajou_nice/with_pet/service/ChatService.java
@@ -8,11 +8,11 @@ import com.ajou_nice.with_pet.domain.dto.chat.ChatRoomRequest;
 import com.ajou_nice.with_pet.domain.dto.chat.ChatRoomResponse;
 import com.ajou_nice.with_pet.domain.entity.ChatMessage;
 import com.ajou_nice.with_pet.domain.entity.ChatRoom;
-import com.ajou_nice.with_pet.domain.entity.PetSitter;
+import com.ajou_nice.with_pet.petsitter.model.entity.PetSitter;
 import com.ajou_nice.with_pet.domain.entity.User;
 import com.ajou_nice.with_pet.repository.ChatMessageRepository;
 import com.ajou_nice.with_pet.repository.ChatRoomRepository;
-import com.ajou_nice.with_pet.repository.PetSitterRepository;
+import com.ajou_nice.with_pet.petsitter.repository.PetSitterRepository;
 import com.ajou_nice.with_pet.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.util.List;
