+++ src/test/java/com/ajou_nice/with_pet/petsitter/main/PetSitterMainControllerTest.java
@@ -8,8 +8,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.ajou_nice.with_pet.CommonApiTest;
-import com.ajou_nice.with_pet.controller.PetSitterController;
-import com.ajou_nice.with_pet.service.PetSitterService;
+import com.ajou_nice.with_pet.petsitter.controller.PetSitterController;
+import com.ajou_nice.with_pet.petsitter.service.PetSitterService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
